requirements.txt
streamlit==1.35.0
simple-salesforce==1.12.5
pandas==2.2.2
plotly==5.22.0
python-dotenv==1.0.1
